compute the list of punctuation characters only once.
The values from `unicodedata` should never change, so this will be a
constant value. On my machine this iterates over 1,114,112 elements to
compute the list. We don't need to recompute this every time and can
save the results of the punctuation characters so only the first time
a RemovePunctuation transform will take time and future calls will
reuse the character list.